pyproject.toml
[project]
name = "auto-doc-agent"
version = "0.1.0"
description = "Automated documentation generation for GitHub repositories using Claude via OpenRouter"
requires-python = ">=3.12"
dependencies = [
    "openhands-ai>=0.9.0",
    "gitpython>=3.1.40",
    "python-dotenv>=1.0.0",
    "litellm>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0",
]
